Please provide the file changes (diffs) for me to generate a commit message.

+++ admin/ajax_files/employees_ajax.php
@@ -1,16 +1,17 @@
 <?php include '../connect.php'; ?>
 <?php include '../Includes/functions/functions.php'; ?>
+<?php include '../../Includes/tenant_context.php'; ?>
 
 
 <?php
-	
 
-	if(isset($_POST['do']) && $_POST['do'] == "Delete")
-	{
-		$employee_id = $_POST['employee_id'];
+$tenant_id = getCurrentTenantId($con);
 
-        $stmt = $con->prepare("DELETE from employees where employee_id = ?");
-        $stmt->execute(array($employee_id));    
-	}
-	
+if (isset($_POST['do']) && $_POST['do'] == "Delete") {
+	$employee_id = $_POST['employee_id'];
+
+	$stmt = $con->prepare("DELETE from employees where employee_id = ? AND tenant_id = ?");
+	$stmt->execute(array($employee_id, $tenant_id));
+}
+
 ?>
